Cleanup

+++ examples/d3d11_tex_consumer.c
@@ -327,7 +327,7 @@ int main(int argc, char **argv) {
 
   texlink_d3d11_texture_frame_t *texture_frames[MAX_FRAMES] = {0};
   ID3D12Fence *sync_fences[MAX_FRAMES] = {0};
-  for (uint32_t i = 0; i < frame_count; i++) {
+  for (int i = 0; i < frame_count; i++) {
     texlink_frame_t *source_frame = texlink_client_frame(client, i);
     texture_frames[i] =
         texlink_d3d11_texture_frame_import(&(texlink_d3d11_import_desc_t){
@@ -391,7 +391,7 @@ int main(int argc, char **argv) {
     }
 
     int idx = texlink_frame_index(frame);
-    if (idx < 0 || (uint32_t)idx >= frame_count)
+    if (idx < 0 || idx >= frame_count)
       idx = 0;
 
     if (wait_shared_fence(sync_fences[idx], texlink_frame_sync_value(frame)) !=
@@ -408,9 +408,9 @@ int main(int argc, char **argv) {
   }
 
   release_window_target(&target);
-  for (uint32_t i = 0; i < frame_count; i++)
+  for (int i = 0; i < frame_count; i++)
     texlink_d3d11_texture_frame_destroy(texture_frames[i]);
-  for (uint32_t i = 0; i < frame_count; i++) {
+  for (int i = 0; i < frame_count; i++) {
     if (sync_fences[i])
       sync_fences[i]->lpVtbl->Release(sync_fences[i]);
   }
